Moved the PluginArea component to the components dir
